ngrams method is working; hmms added (extra method). Flow data discretization needs cleanup

+++ lab3/helper_functions/flow_data_discretization.py
@@ -28,4 +28,4 @@ def extract_code(values):
         for j in range(values.shape[1]):
             code[i] = code[i] + values[i,j] * s/M[j]
             s = s / M[j]
-    return code, M
+    return code.astype(int), M
